requirements.txt
requests
azure-identity
python-dotnet